# perf: Improve launch performance for users with lots of tracks. (1/n)