--- a/mobile/src/api/artist.ts
+++ b/mobile/src/api/artist.ts
@@ -1,34 +1,45 @@
 import { and, eq } from "drizzle-orm";
 
 import { db } from "~/db";
-import type { ArtistWithTracks } from "~/db/schema";
+import type { Artist, ArtistWithTracks, Track } from "~/db/schema";
 import { artists } from "~/db/schema";
 
 import i18next from "~/modules/i18n";
 
 import { iAsc } from "~/lib/drizzle";
 import { deleteImage } from "~/lib/file-system";
-import type { DrizzleFilter, QuerySingleFn } from "./types";
+import type { DrizzleFilter } from "./types";
+import type {
+  QueryManyWithTracksResult,
+  QueryOneWithTracksResult,
+} from "./utils";
+import { getColumns } from "./utils";
 
 //#region GET Methods
-/** Get specified artist. Throws error by default if nothing is found. */
-// @ts-expect-error - Function overloading typing issues [ts(2322)]
-export const getArtist: QuerySingleFn<ArtistWithTracks> = async (
-  id,
-  shouldThrow = true,
-) => {
+/** Get specified artist. Throws error if nothing is found. */
+export async function getArtist<
+  DCols extends keyof Artist,
+  TCols extends keyof Track,
+>(id: string, options?: { columns?: DCols[]; trackColumns?: TCols[] }) {
   const artist = await db.query.artists.findFirst({
     where: eq(artists.name, id),
+    columns: getColumns(options?.columns),
     with: {
       tracks: {
+        columns: getColumns(options?.trackColumns),
         with: { album: true },
         orderBy: (fields) => iAsc(fields.name),
       },
     },
   });
-  if (shouldThrow && !artist) throw new Error(i18next.t("response.noArtists"));
-  return artist;
-};
+  if (!artist) throw new Error(i18next.t("response.noArtists"));
+  return artist as QueryOneWithTracksResult<
+    ArtistWithTracks,
+    DCols,
+    TCols | "album",
+    true
+  >;
+}
 
 /** Get the albums an artist has released in descending order. */
 export async function getArtistAlbums(id: string) {
@@ -39,17 +50,28 @@ export async function getArtistAlbums(id: string) {
 }
 
 /** Get multiple artists. */
-export async function getArtists(where: DrizzleFilter = []) {
+export async function getArtists<
+  DCols extends keyof Artist,
+  TCols extends keyof Track,
+>(options?: {
+  where?: DrizzleFilter;
+  columns?: DCols[];
+  trackColumns?: TCols[];
+}) {
   return db.query.artists.findMany({
-    where: and(...where),
+    where: and(...(options?.where ?? [])),
+    columns: getColumns(options?.columns),
     with: {
       tracks: {
+        columns: getColumns(options?.trackColumns),
         with: { album: true },
         orderBy: (fields) => iAsc(fields.name),
       },
     },
     orderBy: (fields) => iAsc(fields.name),
-  });
+  }) as Promise<
+    QueryManyWithTracksResult<ArtistWithTracks, DCols, TCols | "album", true>
+  >;
 }
 //#endregion
 
@@ -66,7 +88,10 @@ export async function updateArtist(
   id: string,
   values: Partial<typeof artists.$inferInsert>,
 ) {
-  const oldValue = await getArtist(id);
+  const oldValue = await getArtist(id, {
+    columns: ["artwork"],
+    trackColumns: [],
+  });
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const { name: _, ...rest } = values;
   return db.transaction(async (tx) => {