… docs/how-are-tracks-indexed.md
… mobile/src/api/utils.ts
… mobile/src/modules/media/helpers/data.ts
… mobile/src/modules/media/services/RecentList.ts
… mobile/src/modules/media/services/SortPreferences.ts
… mobile/src/modules/scanning/helpers/artwork.ts
… mobile/src/modules/scanning/helpers/audio.ts
… mobile/src/modules/scanning/helpers/rescan.ts
… mobile/src/queries/keyStore.ts
… mobile/src/screens/Sheets/Backup/data.ts
… mobile/src/utils/types.ts