# Why

<!--
Please describe the motivation for this PR, and link to relevant GitHub issues, discussions, or feature requests.
-->

> [!IMPORTANT]
> This won't land in the `v2.1.0` release. This will probably end up as `v2.1.1` or `v2.2.0` due to having a lot of things to change and verify to ensure the performance fixes doesn't break anything.

## Preface
Some users have reported experiencing slow app launches. One commonality is that they have lots of tracks.

Since I typically have around 235 tracks, I ended up duplicating most of them 6 times to get 1481 tracks. At that point, I can notice the slow app launch. There's a couple of contributing factors:
- Unoptimized SQL queries (returning more data than needed).
- Hydrating the Zustand store with data (the Recent List one being a huge contributor as we may be loading many lists).
- The subscriptions in the Zustand store fired after the stores hydrate.
- Logic to validate and save any new tracks (as we need to check every single track to see if it changed).
- Having the "All Tracks" list loaded (as we load every single track, which will be a long time if we have a lot of tracks).
- All the queries that fire once we get out of the onboarding/loading screen.

## Potential Solutions
One solution we're currently working on is updating our API code so that we can limit the scope of the columns returned. From that, we can go to where we use it and specify the columns we need, which will result in less data, but also reduce the time required to get and send that data over.

Since this will be prevalent with large lists, we'll probably end up updating the playback store so that we don't load all the tracks at once (maybe load ±2 tracks around the track at `listIdx`).

Also since the "Recent List" is one of the huge contributors to the problem, we'll probably end up limiting the number of entries displayed.

## Progress

> [!NOTE]
> These numbers will be based around having 1481 tracks, with 5 items in the "Recent List": 2 artists, 1 album, "All Tracks", and "Favorite Tracks". In addition, we have nothing in the playback store (ie: on launch, all we see is the navigation bar - no miniplayer).
>  - It takes ~1.85s to set up the Zustand store (remember nothing is in the playback store).
>    - If we were last playing the "All Tracks" list, this jumps to ~2.76s.
>  - It takes ~1.96s to finish the logic that validates & saves any new tracks (along with cleanup).
>
> We currently haven't taken any measurements on the "lag" caused by all the queries firing once we leave the onboarding/loading screen.
>  - There's probably other things that are unaccounted, but this gives a good starting point for improvement.

### Optimizing the Recent List Store Queries

We applied column limiting on the results with our updated API functions which enables this and observed:
- The ~1.85s drop down to ~0.45s (when we don't have anything in the playback store).
- The ~2.76s drop down to ~1.3s (when the "All Tracks" list is in the playback store).
- If we don't fetch the data for the "Recently Played" list, the hydration time drops to ~0.35s.

Deferring this logic to when we reach the home screen isn't worth it (it actually ends up taking longer).

### Optimizing the Onboarding/Loading Logic

Before we get into the app, we need to check to see if there are any new tracks or tracks that have been modified. This will end up as a **"fixed minimum time"** on the home screen which penalizes users with lots of tracks. Luckily, this doesn't seem to be linear, so it shouldn't be that timely.
- **285 Tracks (Simulated):** ~0.189s to get [basic file information](https://docs.expo.dev/versions/v51.0.0/sdk/media-library/#asset)
- **1481 Tracks:** ~0.581s to get basic file information

This doesn't include the time it takes to validate if any tracks were modified, but it's around an extra 50% of the times listed above.

There's a couple of things we do while on the onboarding/loading screen: checking to see if we need to do any data migration, checking for new or updated tracks, cleaning up the database (if albums or artists are no longer needed), checking for artwork, and deleting unused artwork). Within these chunks, we utilized unoptimized queries (ie: getting all tracks and albums without limiting columns).
- After limiting the columns, we saw the ~1.96s drop down to ~1.3s.

## Summary

In this PR, we simulated a "large" library (to get the feel of a slow startup) with:
- 1481 tracks
- A recent list with 1 album, 2 artists, "All Tracks", and "Favorite Tracks"
- Nothing in the playback store (ie: no miniplayer will be displayed on launch).

We made the following changes:
- Updated the API design of our reusable database table query functions, allowing us to specify the columns returns and whether we want to return the `tracks` relations or the `album` relation with each track.
  - This allows us to reduce the data returned as we don't always need all the fields.
  - **Note:** The option to not return the `tracks` relations with playlists isn't supported (due to different implementation on how we get it).
- Revert decision to allow "get single item" API functions to not throw an error if no item is found (ie: back to original behavior where finding nothing will throw an error).
  - This is to make the typing out the output easier.
- Limit the "Recent List" to display the last 15 lists played.
- Have the following improvements:
  - Reduced store hydration time from ~1.85s to ~0.45s (most of it was from having "All Tracks" in the recent list).
  - Reduced time for doing the checks while on the onboarding/loading screen from ~1.96s to ~1.3s.

# Testing Plan

<!--
Please describe how you tested this change and how a reviewer could reproduce your test, especially if this PR does not include automated tests! If possible, please also provide terminal output and/or screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] Documentation is up to date to reflect these changes.
- [ ] Ensure dependency licenses are up-to-date by running `pnpm sync:licenses`.
- [x] This diff will work correctly for `pnpm android:prod`.